Cloud Functions/MLP-BERT-Classifier/requirements.txt
numpy
google-cloud-storage
scikit-learn
pandas
spotipy